(REF) APIv4 FieldSpec - Extract GuiSpecTrait (label, inputType, inputAttrs, helpPre, helpPost)
